GUI Functionality with Opinionated file handling Got all the GUI components working as to satisfy the following issues: https://github.com/JohnFunkCode/zTournamentSorter/issues/30 https://github.com/JohnFunkCode/zTournamentSorter/issues/31 https://github.com/JohnFunkCode/zTournamentSorter/issues/32 https://github.com/JohnFunkCode/zTournamentSorter/issues/33
In addition to having a fully funcitoning GUI, This updat includes opinioned file handling to put all the files for an instance of a tournament into a rational file folder structure.  This will take file handling out of the hands of the user and will cut down on some of the problems I've seen from poor file managment practices.