Added feature to balance ring splits automaticly
Create a way to split rings into even sized partitions based on the first letter of the last name. The purpose of splitting on the first letter of the last name is so we can tell people who's last name starts with 'X' to go to a specific ring. For example a split could be A-M, N-Z. However, it is not ok to have some people with the last name starting with M go to two different rings. This implies that the partitions may not be exactly even, but they should be as close to even as possible given the split on the first letter of the last name.
See the issue at:
https://github.com/JohnFunkCode/zTournamentSorter/issues/50